From which university did he receive his doctorate?
corrected IVERB.  Using verbnet to allow verb matches.  Using nextCompoundPartObject with verbs for match.  Corrected sentenceEnd. during resolveSpeakers.  Corrected correctSRI in the case of questions.  Corrected Java to read verbSense with new type.

+++ show/src/show/WordMatch.java
@@ -101,7 +101,7 @@ public class WordMatch {
 	short maxLACMatch;
 	short maxLACAACMatch;
 	long objectRole;
-	byte verbSense;
+	int verbSense;
 	byte timeColor;
 	int beginPEMAPosition;
 	int endPEMAPosition;
@@ -189,7 +189,7 @@ public class WordMatch {
 		maxLACMatch = rs.readShort();
 		maxLACAACMatch = rs.readShort();
 		objectRole = rs.readLong();
-		verbSense = rs.readByte();
+		verbSense = rs.readInteger();
 		timeColor = rs.readByte();
 		beginPEMAPosition = rs.readInteger();
 		endPEMAPosition = rs.readInteger();
